+++ src/main/java/com/application/AddressBookAPP/service/AddressBookInterface.java
@@ -4,10 +4,12 @@ import java.util.List;
 
 import com.application.AddressBookAPP.dto.ContactDTO;
 import com.application.AddressBookAPP.model.ContactData;
-
+/**
+ * Service layer Interface
+ */
 public interface AddressBookInterface {
     List<ContactData> getContactData();
-    ContactData getContactDataById(int contactID);
+    ContactData getContactDataById(Long contactID);
     ContactData createContactData(ContactDTO contactDTO);
     ContactData updateContactData(int contactID, ContactDTO contactDTO);
     void deleteContactData(int contactID);
